Simplify resource options
