+++ sdk/dotnet/Pulumi/Resources/ResourceOptions_Copy.cs
@@ -22,12 +22,9 @@ namespace Pulumi
                 Version = options.Version
             };
 
-        internal static ResourceOptions CreateResourceOptionsCopy(ResourceOptions options)
-            => CreateCopy<ResourceOptions>(options);
-
         internal static CustomResourceOptions CreateCustomResourceOptionsCopy(ResourceOptions? options)
         {
-            var copy = CreateCopy<CustomResourceOptions>(options ?? new ResourceOptions());
+            var copy = CreateCopy<CustomResourceOptions>(options ?? new CustomResourceOptions());
 
             var customOptions = options as CustomResourceOptions;
             copy.AdditionalSecretOutputs = customOptions?.AdditionalSecretOutputs.ToList() ?? new List<string>();
@@ -39,7 +36,7 @@ namespace Pulumi
 
         internal static ComponentResourceOptions CreateComponentResourceOptionsCopy(ResourceOptions? options)
         {
-            var copy = CreateCopy<ComponentResourceOptions>(options ?? new ResourceOptions());
+            var copy = CreateCopy<ComponentResourceOptions>(options ?? new ComponentResourceOptions());
 
             var componentOptions = options as ComponentResourceOptions;
             copy.Providers = componentOptions?.Providers.ToList() ?? new List<ProviderResource>();
